requirement.txt
Flask
tensorflow
numpy
pillow
h5py
Werkzeug
flask-cors
scikit-learn
matplotlib